All move around debugger files a bit
